requirements.txt
ultralytics>=8.0.0
torch>=1.7.0
torchvision>=0.8.1
pyyaml>=5.3.1
numpy>=1.18.5
opencv-python>=4.1.2
pillow>=7.1.2
matplotlib>=3.2.2
seaborn>=0.11.0
tqdm>=4.41.0
PyQt5>=5.15.0 